third_party/blink/web_tests/platform/generic/fast/forms/input-step-as-double-expected.txt

Test for a capturing spec change to consider .2 as valid real number for step

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS numberInput.value is "0.5"
PASS successfullyParsed is true

TEST COMPLETE
